refactor: simplify hooks, types, and add comprehensive flow diagrams
## Hook Renaming (Better Clarity)
- Rename useMcpApp → useAppHost (more descriptive, distinct from useMcpApps)
- Rename useMcp.ts → use-mcp.ts (React & Vue - consistent naming)
- Update all imports and exports across codebase

## Type Simplification (Conventional Naming)
- Rename UseMcpAppIframeProps → McpAppIframeProps (remove verbose prefix)
- Rename UseMcpAppIframeReturn → McpAppIframeResult (internal only, not exported)
- Follow React community conventions for type naming

## Documentation Updates
- Update docs/mcp-apps.md with new simplified API examples
- Add 7-phase sequence diagram showing complete MCP app rendering flow
- Update docs/agui-subscriber.md with AG-UI subscriber flow diagram
- Convert ASCII art to Mermaid diagrams for better visualization
- Add component comparison tables explaining differences between hooks
- Update all code examples to use new hook names and patterns

## Flow Diagrams Added
1. Complete MCP Apps Rendering Flow (7 phases):
   - Setup & Discovery → Hook Setup → Tool Execution
   - Component Rendering → App Launch → Data Communication
   - Bidirectional Communication

2. AG-UI Subscriber Pattern Flow (4 phases):
   - Setup → Tool Execution → Event Propagation
   - Tool Call Lifecycle Tracking

## CSS Improvements
- Enhance sidebar caret icon hiding with comprehensive selectors
- Target all possible Docusaurus caret elements for complete removal
