+++ docs/src/css/custom.css
@@ -281,9 +281,15 @@
   margin: 0;
 }
 
-/* --- Improved Sidebar Caret --- */
-.menu__caret {
+/* --- Hide Sidebar Caret Icons --- */
+.menu__caret,
+.menu__link--sublist-caret,
+.menu__caret::before,
+.menu__list-item-collapsible .menu__caret {
   display: none !important;
+  visibility: hidden !important;
+  width: 0 !important;
+  height: 0 !important;
 }
 
 /* --- Sidebar Framework Icons --- */
